requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
requests>=2.31.0
urllib3>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
whitenoise>=6.5.0
python-dateutil>=2.8.2
pytz>=2023.3
